+++ src/main/Carcere_di_gotham_Oggetti_dentro_Oggetti.java
@@ -1,10 +1,10 @@
-package carcere_di_Gotham_FileReader;
+package src.main;
 
 public class Carcere_di_gotham_Oggetti_dentro_Oggetti {
 
 	public static void main(String[] args) {
 		Archivio1 archivio = new Archivio1();
-		
+
 		// dati anagrafici
 		
 		//Detenuti1 markFisher = new Detenuti1("Mark", "Fisher", "13/7/67", "Gotham", "Gotham");
