+++ src/main/Write.java
@@ -1,4 +1,4 @@
-package carcere_di_Gotham_FileReader;
+package src.main;
 
 import java.io.FileNotFoundException;
 //import java.io.FileOutputStream;
